requirements.txt
# Core dependencies
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-multipart==0.0.19
starlette==0.41.3
httptools==0.7.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1

# Agent framework & LangChain ecosystem
deepagents==0.2.4

# LangChain stack
langchain==1.0.4
langchain-core==1.0.3
langchain-community==0.4.1
langchain-classic==1.0.0
langchain-text-splitters==1.0.0
langgraph==1.0.2
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.41

# LLM provider modules
langchain-anthropic==1.0.1
langchain-openai==1.0.2
langchain-ollama==1.0.0
langchain-google-genai==2.1.12
anthropic==0.72.0
openai==2.7.1
ollama==0.6.0

# Google AI
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-auth==2.41.1
googleapis-common-protos==1.72.0
proto-plus==1.26.1
protobuf==6.33.0

# Free web search
duckduckgo-search==7.1.0
ddgs==9.8.0

# MCP Integration
langchain-mcp-adapters==0.1.12
mcp==1.21.0

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0
click==8.3.0
typer-slim==0.20.0
cyclopts==4.2.2
shellingham==1.5.4

# Async & HTTP
aiohttp==3.13.2
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
httpx==0.28.1
httpx-sse==0.4.3
httpcore==1.0.9
h11==0.16.0
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0
anyio==4.11.0
sniffio==1.3.1

# Google Sheets Integration
google-auth-oauthlib==1.2.3
google-auth-httplib2==0.2.1
google-api-python-client==2.187.0
oauthlib==3.3.1
requests-oauthlib==2.0.0
uritemplate==4.2.0

# Gradio UI (compatible with aiofiles 24.1.0)
gradio==5.49.1
gradio_client==1.13.3
ffmpy==1.0.0
pydub==0.25.1
pillow==11.3.0
semantic-version==2.10.0
tomlkit==0.13.3
ruff==0.14.5
safehttpx==0.1.7

# Web scraping & parsing
beautifulsoup4==4.14.2
lxml==6.0.2
soupsieve==2.8
playwright==1.56.0
pyee==13.0.0
greenlet==3.2.4

# Data handling
numpy==2.3.4
pandas==2.3.3
python-dateutil==2.9.0.post0
tzdata==2025.2

# JSON & serialization
orjson==3.11.4
ormsgpack==1.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jsonschema-path==0.3.4

# Markdown & text processing
markdown-it-py==4.0.0
mdurl==0.1.2
markdownify==1.2.0
Pygments==2.19.2
rich==14.2.0
rich-rst==1.3.2

# Tokenization
tiktoken==0.12.0
tokenizers==0.22.1

# Caching & storage
diskcache==5.6.3
cachetools==6.2.1
SQLAlchemy==2.0.44

# Compression
brotli==1.2.0
zstandard==0.25.0

# Cryptography & security
cryptography==46.0.3
PyJWT==2.10.1
Authlib==1.6.5
cffi==2.0.0
pycparser==2.23

# File handling
filelock==3.20.0
filetype==1.2.0
pathvalidate==3.3.1
pathable==0.4.4

# Networking & requests
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
urllib3==2.5.0
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
socksio==1.0.0
dnspython==2.8.0

# XML & SOAP
zeep==4.3.2
isodate==0.7.2
lxml==6.0.2

# Salesforce
simple-salesforce==1.12.9

# Utilities
tqdm==4.67.1
tenacity==9.1.2
regex==2025.11.3
attrs==25.4.0
annotated-types==0.7.0
typing_extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2

# Data structures
frozenlist==1.8.0
multidict==6.7.0
propcache==0.4.1
yarl==1.22.0

# Misc
pyperclip==1.11.0
pytz==2025.2
six==1.17.0
packaging==25.0
distro==1.9.0
platformdirs==4.5.0

# Hashing
xxhash==3.6.0

# Hugging Face
huggingface_hub==1.1.2
fsspec==2025.10.0

# Jinja templating
Jinja2==3.1.6
MarkupSafe==3.0.3

# Dataclasses
dataclasses-json==0.6.7
marshmallow==3.26.1

# Documentation
docstring_parser==0.17.0
docutils==0.22.3

# Key-value storage
py-key-value-aio==0.2.8
py-key-value-shared==0.2.8

# Streaming
sse-starlette==3.0.3

# Other
beartype==0.22.5
bracex==2.6
cloudpickle==3.1.2
fastuuid==0.14.0
primp==0.15.0
referencing==0.36.2
rpds-py==0.28.0
wcmatch==10.1
zipp==3.23.0
jiter==0.12.0
mypy_extensions==1.1.0
annotated-doc==0.0.3
email-validator==2.3.0
hf-xet==1.2.0
litellm==1.79.3
openapi-pydantic==0.5.1

# System keyring (optional, for secure credential storage)
keyring==25.6.0
SecretStorage==3.4.0
jeepney==0.9.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
more-itertools==10.8.0

# ASN1 & RSA
pyasn1==0.6.1
pyasn1_modules==0.4.2
rsa==4.9.1

# gRPC
grpcio==1.76.0
grpcio-status==1.76.0

# UV package manager
uv==0.9.8

# PyYAML
PyYAML==6.0.3

# Metadata
importlib_metadata==8.7.0